+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Core: Update Boot Bios
 - CRT: Renamed a function: `vgs_copy_ptn` -> `vgs_ptn_copy` **(Disruptive)**
+- CRT: Change `vgs_put_bg` to an inline function.
 
 ## Version 0.16.0